+++ js/ui/validation_message.d.ts
